add testing for nmea sentence builder
