+++ prj.conf
@@ -1,2 +1,6 @@
 # KConfig stuff
-CONFIG_STDOUT_CONSOLE=y
+CONFIG_TEST=y
+CONFIG_ZTEST=y
+
+CONFIG_CBPRINTF_FP_SUPPORT=y
+CONFIG_CBPRINTF_LIBC_SUBSTS=y
